+++ src/main/java/com/elektronika/costcalc/model/MaterialCost.java
@@ -2,15 +2,16 @@ package com.elektronika.costcalc.model;
 
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @DiscriminatorValue("materialCost")
 public class MaterialCost extends Cost {
 
 
-    private Float selfProducedMaterialCost;
+    private BigDecimal selfProducedMaterialCost;
 
-    public MaterialCost(Product productId, Float cost, Integer yearMonth, Float selfProducedMaterialCost, boolean directCost) {
+    public MaterialCost(Product productId, BigDecimal cost, Integer yearMonth, BigDecimal selfProducedMaterialCost, boolean directCost) {
         super(productId, cost, yearMonth, directCost);
         this.selfProducedMaterialCost = selfProducedMaterialCost;
     }
@@ -19,11 +20,11 @@ public class MaterialCost extends Cost {
     }
 
 
-    public Float getSelfProducedMaterialCost() {
+    public BigDecimal getSelfProducedMaterialCost() {
         return selfProducedMaterialCost;
     }
 
-    public void setSelfProducedMaterialCost(Float selfProducedMaterialCost) {
+    public void setSelfProducedMaterialCost(BigDecimal selfProducedMaterialCost) {
         this.selfProducedMaterialCost = selfProducedMaterialCost;
     }
